chore: move expression parsing utilities
